Update seeds for current year
Previously, we had to add a new line for `AnnualSchedule` for each year. Since from 2018 onward we've used the same set of fields, just different dates, I added a loop to the seeds from 2018 to whatever the current year is and it will add an `AnnualSchedule` for each year. I also went ahead and added a `.ruby-gemset` file to create a gemset if the user has rbenv or rvm installed so we won't have any conflicts with other projects.